+++ streamlit_app/script_pages/page_shots_calc_integrals.py
@@ -64,7 +64,7 @@ def read_dataframe(file):
             break
     #then read the data frame
     df = pd.read_csv(file, skiprows = lines_to_skip)
-    return df, capacitor, int(integration_time)
+    return df, capacitor, float(integration_time)
 
 df, capacitor, inegration_time_us = read_dataframe(os.path.join("Measurements", "Shots", filename))
